Rename package from com.diffblue to io.diffblue
Fuzztest does not work properly on projects in com.diffblue.
This is a bug in fuzztest, but the workaround here is straightforward.
