--- src/main/java/com/diffblue/corebanking/CoreBankingException.java
@@ -1,13 +1,0 @@
-package com.diffblue.corebanking;
-
-/** CoreBankingException. */
-public class CoreBankingException extends Exception {
-  /**
-   * Constructor.
-   *
-   * @param errorMessage The exception error message.
-   */
-  public CoreBankingException(String errorMessage) {
-    super(errorMessage);
-  }
-}
