--- src/main/java/com/diffblue/corebanking/account/AccountException.java
@@ -1,15 +1,0 @@
-package com.diffblue.corebanking.account;
-
-import com.diffblue.corebanking.CoreBankingException;
-
-/** AccountStatement exception. */
-public class AccountException extends CoreBankingException {
-  /**
-   * The constructor.
-   *
-   * @param errorMesssage The exception error message.
-   */
-  public AccountException(String errorMesssage) {
-    super(errorMesssage);
-  }
-}
