--- src/main/java/com/diffblue/corebanking/datamanagement/WriteToDB.java
@@ -1,9 +1,0 @@
-package com.diffblue.corebanking.datamanagement;
-
-import com.diffblue.corebanking.CoreBanking;
-
-public class WriteToDB {
-    public static void writeToDisk(CoreBanking coreBanking) {
-
-    }
-}
