+++ src/main/java/io/diffblue/corebanking/datamanagement/WriteToDB.java
@@ -1,0 +1,9 @@
+package io.diffblue.corebanking.datamanagement;
+
+import io.diffblue.corebanking.CoreBanking;
+
+public class WriteToDB {
+    public static void writeToDisk(CoreBanking coreBanking) {
+
+    }
+}
