+++ src/main/java/io/diffblue/corebanking/transaction/TransactionException.java
@@ -1,0 +1,14 @@
+package io.diffblue.corebanking.transaction;
+
+/** TransactionException class. */
+public class TransactionException extends Exception {
+
+  /**
+   * TransactionException constructor.
+   *
+   * @param errorMsg The error message of the Exception.
+   */
+  public TransactionException(String errorMsg) {
+    super(errorMsg);
+  }
+}
